feat: Add `output_path` property, and streamline implementation
Users got confused about the semantics of the `output_path_prefix`
property. This patch makes it better by renaming it to `output_path`,
while still providing backwards-compatibility.

Further, the `destination_path` property can also be used as an alias,
thus this target can be a drop-in replacement to the `hotgluexyz`
variant.

The value of the variable does not need to be configured using a
trailing slash any longer. Instead, the implementation more thoroughly
leverages `pathlib.Path` for concatenating `output_path` and `filename`.