feat: Add `output_path` property, and streamline implementation (#50)
## About

Users got confused about the semantics of the `output_path_prefix`
property. This patch makes it better by renaming it to `output_path`,
while still providing backwards-compatibility.

Further, the `destination_path` property can also be used, thus this
target can be a drop-in replacement to the `hotgluexyz` variant.

The value of the variable does not need to be configured using a
trailing slash any longer. Instead, the implementation more thoroughly
leverages `pathlib.Path` for concatenating `output_path` and `filename`.

## References

- GH-9

## Thoughts

I am not sure about fc54324285f5. I added it because mypy tripped like
described in the commit message. Please educate me if that was wrong, so
I will either remove it or amend it correspondingly.